Don't ship td.c or vhd.h
After discussing with Andy W we might want to do exciting things in td.c in the
future so keep it private.

vhd.h is now unused in the shipped source so don't ship that either.

That just leaves profile.h as code which is proprietary but shipped. We don't
think there's anything secret in there.

Signed-off-by: Ian Campbell <ian.campbell@xensource.com>