Remove dependency on the presence of a Xen repository by copying the build system infrastructure which blktap uses from the Xen repository into the blktap repository. Partly inspired because the dependency was a bit bogus but it also happens to make it possible to build the blktap component in both the trunk and orlando-newxen trees without hacks.
This initial import is a simple straight copy with
s/XEN_ROOT/BLKTAP_ROOT/g there are large section of the new files
which make no sense in the blktap context and can probably be dropped.

The XEN_ROOT variable is still required if USE_SYSTEM_LIBRARIES != y
in order to pickup the libaio code.

Confirmed by comparing the build logs that the only difference is that
XEN_ROOT= is no longer passed to make, all compiler options etc remain
the same.