Refactor libvhd::vhd_open, replacing generic O_FLAGS with vhd-specific flags.
New vhd-specific flags include VHD_OPEN_FAST and VHD_OPEN_STRICT.

VHD_OPEN_FAST optimizes vhd_open by reading the backup footer at the beginning
of the file rather than the primary footer at the end of the file; typically,
this allows both the header and footer to be read with one read().

VHD_OPEN_STRICT, on the other hand, will only try to read the footer at the end
of the file and will return a failure if the footer is not valid; this is
useful for operations that should only be run against inactive vhds.