CA-24604: Degrade duplicate pause events to a debug level warning.
Changeset 379:821c0729450c provokes bogus events due to a short race
window between (1) registering the pause watch and (3) receiving the
implicit event for (1) , if (2) an actual write precedes the event for
(1). CA-24604 comprises a quicktest run which actually hit this
condition.

This is already non-fatal and should remain in the logs. But it can
not be promoted to a tapdisk-error. [Where the agent presently ignores
it anyways].

CA-24296: And while we are at it, do not wait for the pause to succeed
in order to be able to detect it.