CA-26523: Adjust blktapctrl exit status to main loop result.
In summary, this remains as unused as the prior code: the catch is
that we presently never exit voluntarily.

A watchdog parent + termination on signals, fatal conditions, etc,
would change the landscape. Especially since exiting with running
tapdisks is presently irrecoverable.

So for now just clean stuff up a little to prevent people from having
to investigate where all those return values go.