CA-19010: Make vhd-util repair work on block devices.
The repair facility opens a VHD, recovers metadata from the backup
footer, then writes the data back to the primary footer at
end-of-file.

We only had it implemented for file-based VDIs, the difference being
that our definition of 'end-of-file' is a slightly different one on
block devices.

This patch removes vhd-util's own assumptions about what the right
footer offset is and lets instead vhd_write_footer() do the math. The
price is a redundant backup footer re-write in the repair case and a
redundant ftruncate() issued by tapdisk's vhd_close() in the
non-faulty case.

None of those are on a critical path. So let's avoid to declare more
funny functions non-static.