CA-30816: Fix shutdown phase on device reset.
Cset 397:272912c3d0ec (CA-26523) broke the tapdisk shutdown phase run
on a reset of the backend's XenStore root node. I missed the fact that
this is the point where the shutdown message in channel_close() is
essential.

This patch keeps the channel_reap() operation to close the startup
failure loophole, but doesn't defer channel_close(). We now free the
channel only - but unconditionally - after tapdisk exit.

Removed the comment on safe errors since it doesn't apply any more. At
the expense of splitting the original close code.