CA-34846: Integrate tapdisk-syslog with the tlog_error path.
Currently removes the caller line filter, based on the fact that we
are only reporting terminal vreq failures anyway.

[The alternative would have been to keep filtering and flush the log
only once per loop iteration.]

With an overall request timeout of 2 minutes per request, there should
presently be no need to filter.

This may likely change in future, so bursts of errors would lead to
message loss. We anticipate this by logging to both syslog and the
logfile, which is then reliable.