blktap2: Update tap-ctl timeouts and blocking behavior.
Adds a -t <timeout> switch to the destroy, pause and close operations,
which are known to block. Changes the library call interfaces to use
timeval structs, not <int> secs, which are then passed around
internally. On Linux, the effect should be that we track the total
timeout across calls, not a maximum for individual operations.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>