PR-1053: Fix tapdisk-stream.
Was dropped by the vreq patch, because the VBD change broke
compilation. Essentially acts as a tap, but in a somewhat improperly
(now) because it emulates a blkif ring to do so.

Fixed with a major rewrite. The tool got somewhat out of fashion, but
is a good API exercise:

 * New code now queues vreqs instead of fake blkif reqs.
 * Won't get rid of the loopback event fd yet.

The latter is because new requests are only processed per iteration,
after select fired. As a consequence, vreq completion cannot just
queue new requests, because they won't get run before the next
external event fires. This will affect VBD I/O queued by filter
drivers, too.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>