PR-1129: Fix main loop timeout bug.
Calls to scheduler_check_fd_events return prematurely, failing to mark
a timeout pending, iff nfds is 0.

This leads to a condition where the deadline never gets reset, so the
next calculated timeout tv will be {0, 0} and we're left spinning in
select()s.

Fixed by separating fd and timeout checks.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>