Make blktap portable to Ubuntu
These are changes mostly required by gcc 4.5, with the exception of the change
to Config.mk, which inserts some make variables that we expect to be inserted
by the spec file. The rest of the changes are including missing headers,
initialising a variable and providing a mode when opening a file.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>