CA-86426: fix a tap-ctl race.
It is possible for a new control connection to be accepted, but before anything is
received on it, for a "detach" message to come in and cause the tapdisk server to
exit. On cleanup, the new, unprocessed connection causes this:
tap-err:tapdisk_ctl_conn_drain: tapdisk-control.c:273: FAILED ASSERTION: 'conn->out.done'

We now simply drop any such connections.