CA-110509: Impossible to create LVHD over iSCSI SRs with more then one iSCSI target IP
When passing multiples IPs on the target list, when printing entries for the
wildcard option we compose a string this way
IP1,IP2,..,IPn:PORT
The IPs can be in IPv4 or IPv6 format but the last colon is always
a delimiter for the port.
The parse_IP_port function is not suitable in this case (it needs to be modified, anyway),
so we simply split the PORT from the rest of the string: the latter is kept as it is.