CP-5818: Productise removal of lvm2-lvcreate-inactive-flag.patch
Snapshot creation process uses a temp file with all snapshot information in
the name of the file. Since this name is too long and not liked my mapper,
the file is created in an inactive mode. Changed the logic to create a file
with shorter name and all informaton is written inside the file. Inactive flag
for lvcreate is no longer required.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>

GitHub: closes #31 on xapi-project/sm