CP-5946: remove the sleep hack around xc_evtchn_bind_interdomain
The sleep hack was originally introduced to work around a failure
around xc_evtchn_bind_interdomain. With the new 3.x kernel, I
couldn't reproduce the failure with this hack commented out. I
tried iterations of tests on a few different setups on different
hardware, the result seems persistent. So I'd propose to remove
the hack, and hopefully it should just work. We can investigate
further if any new problem is spot later.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>