CP-6167: Fix some bad practice reported by pylint
We conditionally assign a variable with either an object or a string.
Though this might not cause problems at runtime if we can guarantee
that we'll always branch by inspecting the runtime value, it is a
dangerous practice nonetheless.

Pylint complained with the following error: "Instance of 'str' has no
'xenapi' member".

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>