CP-6167: Work around an issue pylint isn't clever to understand
This is not a bug, it's because pylint can't figure out magics
like runtime types based dispatching with isinstance etc. and
will issue an E1103 error.

We don't want to disable E1103 all together, as this could shadow
some real issues later on, so we work around the issue here. We'd
better refactor our code to implement the same functionality
differently.

Another way to work around this is to annotate the specific
occurrence in the source code with the pylint ignore tag.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>