CA-118742: Drop whitelisting of HBA device drivers from devscan.py
CA-14825 introduced device driver whitelisting when probing HBAs to avoid that
local disks are listed when calling "xe sr-probe type=lvmohba".

Today the use of shared storage using SAS, FC, FCOE and iSCSI is the norm and
it is becoming increasingly difficult to ensure that the whitelisting of device
drivers is current.

Replacing the whitelisting with blacklisting for explicitly local device driver
groups should improve the customer experience, interoperability of SM, reduce
the maintenance effort and at the same time reduce the complexity of the code.

It seems less concerning that users might end up provisioning a local disk, if
it happens to have scsi id, with LVHDoHBA instead of LVHD, than that users end
up not beeing able to provision their shared storage at all. Also the chance of
missing a local device driver appears much lower due to the more consistent
naming.

After all there is the question, whether we should really prevent users from
provisioning local disks using lvhdohba - but changing this is only a 1-line
change, to adapt the blacklist, after this commit.

Signed-off-by: Robert Breker <robert.breker@citrix.com>
Reviewed-by: Vineeth Thampi Raveendran <vineeth.thampi@citrix.com>

GitHub: closes #45 on xapi-project/sm