CA-113323: Do not issue udevadm trigger on SR creation.
Rolled back original fix for CA-105130 which triggers udev events
for all devices in the system when a new SR is created. New events
for existing devices coupled with the ignore_device option in the
event handling rules caused the device to be removed.

The bug addressed in CA-105130 doesn't exist in the new release and
it should be safe to roll back the patch.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Vineeth Thampi Raveendran <vineeth.thampi@citrix.com>