CP-6337: Ensure multipath kernel modules are present
Once started the multipath daemon does not load the needed
kernel modules.
In the past it was not a problem because they are loaded as soon as
a device is added using the CLI.
Now we need some checks before doing this and they fail if this modules
are not loaded.

Loading modules while starting multipath is the standard behaviour in
more recent versions of multipath.
We just ensure this is always true.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Keith Petley <keith.petley@citrix.com>