CP-5818: Productise removal of lvm2-lvcreate-inactive-flag.patch
Snapshot and leaf coalesce operations use journaling files with
additional info like parent UUID in the name of the file. Since the
resultant file name is too long for LVM, the file was created in an
inactive mode. This patch changes the logic to create a journal file
with a shorter name and writes additional informaton into the file.
Inactive flag for lvcreate is no longer required.

This changeset includes the fix made for ticket [CA-117034].

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>

GitHub: closes #50 on xapi-project/sm