CA-121481: No need to use the daemon
The script does not need the daemon to get the needed info.
This script is executed both as a standard boot script and
during the installation ensure the root disk is whitelisted.
In the latter case the daemon is not supposed to be running
but the root disk can be managed by the device mapper.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>