CA-122566: Fixed boot from multipathed SAN
A bug in Centos 6.4 multipath causes unnecessary delays
in root disk partition creation if detect_prio is set to yes
in multipath.conf(default value) and the device doesn't support
it. Setting detect_prio to no provides a workaround for the problem
and allows boot from a multipathed SAN. Bug is fixed upstream
in Centos 6.5 multipath.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Vineeth Thampi Raveendran <vineeth.thampi@citrix.com>