CP-7292: remove dependency on iofi
iofi is a customize module which doesn't exist in either xenbuilder
chroot or a standard XenServer host. Hence pylint complains about
the reference to a non-existed module.

My understanding is that the module is for debugging purpose and can
be purposely copied onto the system when necessary. But in general
cases, it won't be on the system.

The change workaround the pylint complain by using dynamic loading
via module name. This also fits the conditional loading concept
in context, where loading which module could be calculated and
dynamically decided (though not necessary in this instance).

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>