CP-7292: Make use of the exit code of pylint
In the previous version of pylint we use, the exit code of a pylint
process is always zero (regardless there are issues found or not).
So we have to capture the output of pylint to make a judgement call
ourselves.

The newer vesion of pylint exit with non-zero code if any issue
has been found. So we can now interpreter the exit code rathe than
the output text, which makes things simpler.

Unfortunately, a bug in the exact version 0.21.1 [1] stops us from
passing all the python files to check to pylint all at once, so we
still have to iterate over all the files one by one and record the
existence of any errors. The code can become considerablly simpler
when we move to a even newer version of pylint with this bug fixed.

[1] http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=591670

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>