CA-122252: Kill requests if the VBD resume has failed
If a resume operation fails (e.g. the LUN has gone offline) on a
previously paused VBD for which there are pending requests, these
requests will never be completed. This results in blkback not being
able to shutdown as it waits for pending requests to be completed. This
patch addresses this problem by killing all requests in the event of a
failed resume. If the problem for which the resume failed is fixed,
resuming again the VBD should work.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>