CA-120162: don't fail if the GC can't be kicked
When a VDI.delete needs to abort the GC and restart it, a deadlock might
occur which leads to reporting that VDI.delete failed while it has
partially succeeded, as the actual deletion of the VDI will happen
sooner or later. Fixing the deadlock is not straightforward, so the best
effort solution is to simply ignore such failures.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>

GitHub: closes #83 on xapi-project/sm