CA-125489: Wait for the hotplug scripts before initialising
This patch addresses a race condition where the back-end switches to the
InitWait state too soon and the kernel thinks the hotplug script have
run. XAPI waits for these hotplug scripts to run, which never happens,
and the domain is never actually started. We address this issue by
ensuring that the back-end doesn't respond to front-end state changes
unless the hotplug-status XenStore key receives the value "connected".

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>