CA-125917: fix drain queue when closing
tapdisk_control_close_image waits for the Xenbus to disconnect, but this
can never happen because only a tapdisk_control_xenblkif_disconnect
message can do this, and that cannot be executed because the current
tap-ctl close command is still pending. The solution is to have tap-ctl
close to try and disconnect from the ring if necessary.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>