CA-126092: Retry tap-ctl close till it succeeds or times out
If tapdisk is busy processing another command (like disconnect
from ring), an EBUSY status is sent back to the tap-ctl close
command. In this commit, we wait and retry the tap-ctl close
command till tapdisk responds. The wait is timed out at 120 seconds
which is tapdisk's timeout for command completion.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>