CP-7800: mask conflicting commands instead of retrying
The Storage Manager and tapback might concurrently issue exclusive
commands to tapdisk. tapdisk is not designed to handle concurrent
commands and if this occurs it fails the other command with EBUSY. We
address this issue by masking the event associated with the command that
would be rejecting while tapdisk is busy servicing the first command, and
unmasking it when the first command has finished.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>