Revert "CAR-96/CA-7800: mask conflicting commands instead of retrying"
This reverts commit ad30ba6b6d5e99fd011640c517fd2ee0520798e2.

CA-7800 (actually it's CP-7800), doesn't fix the tap-ctl race conditions
as it masks the wrong event. The connection object as is not associated
with an event per se, only the input/output operations that belong to it
are. Fixing this is not trivial, so to save time we revert the hack-ish
timeout-based solution.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>