revert CA-114341 patches
The CA-114341 patches fail to pass the BST as some curious locking
issues appear. It's not yet clear whether these patches are just wrong
or whether they uncover an existing issue. I'm removing them for now to
unblock the BST and the merge into trunk.

This patch reverts the following commits:
* 8c1c069c: "CA-114341: don't try to reacquire the lock"
* 1f5db2b3: "CA-114341: explicitly use keyword argument"
* a47ceebb: "CA-114341: GC: acquire VDI lock non-blocking"

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>

GitHub: closes #99 on xapi-project/sm