CA-126097: Handle vm-clone when VBD is attached.
A clone operation was never performed when the VM is running, as a result
there is no need to pause/unpause tapdisk during the clone operation. With
storage motion in place, there is a change in this behaviour. vm-copy attaches
and activates the VBD, a clone operation during the phase could result
in VHDs in inconsistent state.

Signed-off-by: Vineeth Thampi Raveendran <vineeth.thampi@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>

GitHub: closes #101 on xapi-project/sm