CA-128872: suspend ring operation on VBD pause
Pausing the VBD without suspending the operation of the ring leads to
requests being extracted from the ring but not being submitted as the
VBD is paused. The problem in that behaviour is that in SXM tapdisk is
paused before it is closed, so there are requests trapped in tapdisk.
Suspending the operation of the ring when pausing the VBD ensures the
ring is flushed and no requests are extracted from the ring while the
VBD is paused.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>