CA-135017: notify toolstack about tapdisk unpause errors in coalesce
Coalesce is an asynchronous operation so if a tapdisk refresh fails the
VM will lose the disk but the administrator will realise that when the
VM hangs. This patch notifies the toolstack as soon as the failure
occurs.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Acked-by: Germano Percossi <germano.percossi@citrix.com>
Imported-by: Thanos Makatos <thanos.makatos@tcitrix.com>

(cherry picked from commit 5d3a649127036e0c2ceeeb169d55e451e22cdd91)