[CA-106558] & [CA-94713]: parameterise NFS request timeout value.
Allow the user to specify the maximum time a VBD should wait before reporting
errors in the case of NFS outages. The new parameter is
SR.other-config:nfs-timeout and it is specified as an integer number of
seconds. This parameter will affect the NFS timeo mount option and the tapdisk
request timeout limit.

Signed-off-by: Andrei Lifchits <andrei.lifchits@citrix.com>
[Chandrika Srinivasan imported and modified patch for xs64bit branch]
Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Ross Lagerwall <ross.lagerwall@citrix.com>