CA-137830: enable memory caching for the VHD cache in IntelliCache
This patch enables memory caching for the VHD file acting as the cache
in IntelliCache, despite the fact that it is opened in R/W mode. Without
this, read caching interacts badly with IntelliCahce. This is safe to do
so as the cache file is always thrown away.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
(cherry picked from commit 1a6e65d38f38a61c233eb741f785fc6bcaeeda34)

Reason for cherry picking:
    [CP-9440] Merge read caching related changes to Master