CA-137830: change behaviour of vhd_cache_enabled
vhd_cache_enabled must return true not only if the VHD_CACHED_ENABLED
bit is set, but also if the VHD file in opened in read only mode
(VHD_OPEN_RDWR is not set). This ensures that if the leaf is opened R/W
and for buffered I/O (which is the case for IntelliCache), the parent
chain is is not considered 'cached'.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
(cherry picked from commit 9272233a0e304c9d765503d3dbc43ed40a13f46c)

Reason for cherry picking:
    [CP-9440] Merge read caching related changes to Master