CP-8633: Remove support for MPP RDAC LUNs
This change makes sure that no calls to MPP RDAC drivers (which are not
supported anymore) are made through the storage layer. A check for RDAC LUN
always returns false but not before logging a warning if the RDAC LUN path
is found to be populated.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>
Imported-by: Thanos Makatos <thanos.makatos@citrix.com>

(cherry picked from commit 0f4c99242ed67af6cd203676f2a426980adad456)