CA-136792: produce raw blkback-style stats
Produce I/O statistics in /dev/shm/vbd3-domid-devid. We use one file for
all seven counters as otherwise we're wasting too much memory. Unlike
blkback, the metric are in raw binary format, as this is the most
efficient way for producing them. The format of the statistics is
defined in /usr/include/blktap/blktap3.h.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>

GitHub: closes #97 on xapi-project/blktap