CA-142603: use correct tapdisk log rotation rules
Rotation of logs is performed by /opt/xensource/bin/rotate_logs_by_size
(and not by logrotate itself) and the rules for that are specified in
/etc/logrotate-xenserver.conf. That configuration files contains
contains the wrong rule for blktap (it doesn't send the HUP signal).
Also, we already specify a logrotate configuration file in
/etc/logrotate.d/blktap which was never used.

This patch includes it in /etc/logrotate-xenserver.conf. Also, log
size, log age, and other options must come from
/etc/logrotate-xenserver.conf itself so they are removed from
/etc/logrotate.d/blktap.

Signed-off-by: Thanos Makatos <thanos.makatos@citrix.com>
Reviewed-by: Mate Lakat <mate.lakat@citrix.com>