CP-12399: Prevent tapback from plugging non-blktap2 devices
With the introduction of tapdisk3, the LUN-per-VDI SRs will either fail or
cause the wrong VDI to be plugged to guests. That is because the SM will
create a block device on /dev/sm/backend/<sr_uuid>/<vdi_uuid> that actually
uses the major/minor number of the actual (raw and block) VDI (instead of the
corresponding blktap2 block device as in releases before Creedence).

This patch makes tapback ignore all physical-device entries that do not have
the major number matching the blktap2 major. It therefore prevents VMs from
getting the wrong virtual disk.

Separate commits (to the toolstack and to the SM) will make the SM instruct
the toolstack to write on the .../vbd/... xenstore path (instead of /vbd3/)
for such VDI types.

Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>
Reviewed-by: Stefano Panella <stefano.panella@citrix.com>

(cherry picked from commit 0246ce181a23a88307a8ff442f13199ef27e6afa)
Imported-by: Pritha Srivastava <pritha.srivastava@citrix.com>
Merge: CA-183474