CA-165129: Ensure the mounted CIFS share is at least RO
Windows sharing and security settings can be configured in a way that a share
can be mounted, but not even read is allowed on it. This check ensures that if
this is the case, the user is informed and the share is unmounted.

On the side: move the removal of the credentials file in a finally clause in
the mount.cifs try/except block.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>