CA-148438: Expose mem_mode flag in statistics file
Each tapdisk's memory mode state is now exposed in the respective
'/dev/shm/vbd3-*/statistics' memory mapped file. This will be
picked up by xcp-rrdd-iostat which will then aggregate it for the
total number of tapdisks on the host and expose a single rrd with
the total number of tapdisks in low memory mode.

See also CP-12967.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>