CP-13963: Use proper parameters for clone and snapshot
This is addressing the problems caused with the conversion
from percentage to absolute of xlvhd parameters.

Thanks to this change it is now possible to use an initial
allocation of 0 which is the default for the xlvhd SR.

By design we would like that on clone and snapshot operation
instead of using the original VDI parameters for
initial_allocation and allocation_quantum, we would use the
SR ones.

Being the SR default initial_allocation = 0, we can expect
that every clone or snapshot will use minimal space.

Added a fix for a problem where after the leaf-coalesce
caused by the GC, the leaf was left fully inflated as in the
thick case. Now there is a check so in case of xlvhd we do not
fully inflate the leaf.

Added a fix to convert the allocation_quantum to MBs from bytes
before passing it to tapdisk which is only accepting MBs.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>
Acked-by: Germano Percossi <germano.percossi@citrix.com>