Quick fix for xenvmd conf parsing after RRDs field added
The configuration files for the xenvm family (the one generated
by setvginfo, the one for xenvmd and the one for the local allocator)
need to be parsed using s-expressions and not rely on the line
number.

This commit addresses quickly the problem caused by the addition
of the RRDs field in the configuration file.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>