CA-182713 : lvm.conf file is wrong in some systems
This happens only when we upgrade SM rpm. During an upgrade, after installation
of the new package, old package is removed and %postun is called. In %postun
all changes done to the lvm.conf file gets reset to lvm.conf.orig which is the
expected behaviour during package uninstall and not during package upgrade. As
a fix to this issue a logic is added in %postun to differentiate between package
uninstall and package upgrade. This prevents lvm.conf to be reset during a
package upgrade.

Signed-off-by: Sharath Babu <sharath.babu@citrix.com>
Reviewed-by: Stefano Panella <stefano.panella@citrix.com>

GitHub: closes xapi-project/sm#276