Fixed flag size from uint8_t to uint16_t
We ran out of flags (we have 9 now) so the size of the flag
is not sufficient.
In theory there should not be any problem but it is not clear
if this flag is used to be written somewhere where it is expected to
be 8 bits.

To be checked later

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>