CA-173369: Make thinprovd crash consistent
Now thinprovd, if restarted, will automatically manage the same
VGs it was already managing before the crash.

This has been implemented changing thin-cli to create/delete
a file with the name of the VG just after the VG has been successfully
added/deleted from the server, so there is no race.

The files are created/removed in /var/run/thinprovd/

On a machine just booted, /var/run/thinprovd/ does not exists and is
created by thinprovd when started.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>