CA-174530: Fix udev rules for multipath boot-from-SAN
The problem with the existing rules is that the raw partitions (e.g.
sda1) are not removed before the device-mapper devices are created which
means that for a short time, disk/by-label symlinks exist pointing to a
part of a multipath device. This causes issues.

Instead, use the upstream udev rules which work properly, and use a
separate udev rule to run mpathcount to inform XAPI of path changes.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>