Fixed circular dependency introduced by original Kuiper code
I think the problem should be fixed in lvhdutil.py given
lvutil is a less specialized module and could be called by
lvhdutil.py.
Unfortunately the import in lvutil was introduced long time
ago, so the easisest way of fixing it now is in the import
in lvhdutil which came last.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>