CA-174119: Change the way to determine the SR allocation type
From: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>

This patch completely removes the "querying XAPI about an SR's allocation
type, storing it in an environment variable and read it from there when
needed" functionality. Instead, we rely on the existence of the xenvm
config files in '/etc/xenvm.d' to determine the SR allocation type.
These files are created in SR attach() and removed in SR detach().

Furthermore, the functions that need this information before the creation
of the aforementioned config files, take it explicitely as input, by
extracting it from the SR command.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>