Have tapback announce the discard capability for vbds
The discard capability is announced for vbds if the following two conditions
apply:
* The backing tapdisk indicates discard support for the image that is
  attached to the vbd (i.e. block-aio is used in combination with a capable
  filesystem)
* tapback is not run with the --nodiscard overwrite

If discard is enabled, the discard granularity (discard_granularity) is
statically announced as the sector size of the vbd. The alignment
(discard_alignment) is statically announced as 0.
Secure discard (discard_secure) is never guaranteed.

This commit has been dev-tested using:
* v8 Windows PV drivers that include XenDisk and thereby implement discard
* Linux xen-blkfront that implements discard

Signed-off-by: Robert Breker <robert.breker@citrix.com>