CA-184700: Refine initial_allocation for clone and snapshot
From: Stefano Panella <Stefano.Panella@citrix.com>

Existing code was not very precise in case on thin lvhd snapshot
and clone operation, also the value for initial_allocation
was not saved in sm_config.

Now we differentiate between snapshot and clone operation.

In case of snapshot the new two VDIs created are
1) leaf for existing VM -> ia = SR initial_allocation
2) snapshoot            -> ia = 0

In case of clone the new two VDIs created are:
1) leaf for existing VM -> ia = SR initial_allocation
2) leaf for new VM      -> ia = SR initial_allocation

The initial_allocation for these new VDIs is also saved in
their sm_config.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>
Reviewed-by: Jorge Martin <jorge.martin@citrix.com>