CA-179190: Decrese min host pool to 10 times min aq
From: Stefano Panella <Stefano.Panella@citrix.com>

This is because the min host pool is currently set to 1GiB and
this is a problem when we have LUNs of just 1GB (as the ones we
use for the scalability test of creating 250 SRs).

There is no reason why the minimum should not be lower, so I
have put it to 160MiB which is 10 time the current
min allocation_quantum.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>
Reviewed-by: Jorge Martin <jorge.martin@citrix.com>