CP-12972: scheduler: use microsecond granularity
Previously, the scheduler worked on integer-second granularity. Now, we allow
timeouts and deadlines of fractions of seconds.

This is achieved by using struct timeval rather than int as the type of the
timeout and deadline values.

We retain the notion of -1 meaning infinite timeout. For this the tv_sec member
of the struct timeval is used.

Note that the following functions' parameters still deal in integer seconds:
 - tapdisk_server_set_max_timeout,
 - tapdisk_server_register_event.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Stefano Panella <stefano.panella@citrix.com>