CP-12972: scheduler: refactor timeout arithmetic to use macros
No semantic changes.

This prepares us for a neat changeover to use non-integer timeouts. (In
addition, it hides the "magic" of the infinite timeout value.)

Most of the timeout arithmetic is inside scheduler.c, but two functions exposing
timeouts spill outside:
 - tapdisk_server_event_set_timeout and
 - tapdisk_server_register_event.

Therefore we put the macros in a new header file so they can be accessed from
that function's callers.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>