CP-12971: Add infrastructure to perform blkif ring polling
Add the infrastructure for polling. But don't actually ever enter polling mode
for now.

New function tapdisk_start_polling is the entry point. It reuses the existing
"chkrng" event, scheduling this immediately.

The chkrng event's callback is modified so that when in polling mode, this event
is kept scheduled to fire immediately, and the blkif ring's event counter is not
updated.

tapdisk_start_polling also schedules an event to stop polling (currently
hard-coded to 5 seconds in the future). That event's callback exits polling mode
and sets the ring's event counter.

tapdisk_xenio_ctx_process_ring is modified so that any requests arriving in the
blkif ring during a period of polling pushes the stop-polling event further into
the future.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>