CP-12971: Make polling duration configurable through xenstore
A VBD's polling-duration key in xenstore represents the duration of polling,
expressed in milliseconds. If the key is not present, 0 is assumed, meaning no
polling.

Example path: /local/domain/0/backend/vbd3/1/51712/polling-duration

This value is read by tapback when connecting to the ring, stored as a member of
struct vbd, and passed to tapdisk as a new parameter of the
TAPDISK_MESSAGE_XENBLKIF_CONNECT message.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>